implementing comments on the review, restoring vectorization

+++ include/common/defs.h
@@ -144,8 +144,8 @@ namespace gridtools{  namespace enumtype{
 #ifdef CXX11_ENABLED
     using int_t          = int;
     using short_t        = int;
-    using uint_t         = int;
-    using ushort_t       = int;
+    using uint_t         = long unsigned int;
+    using ushort_t       = unsigned int;
     template<int_t N>
     using  static_int=boost::mpl::integral_c<int_t,N>;
     template<uint_t N>
@@ -155,10 +155,10 @@ namespace gridtools{  namespace enumtype{
     template<ushort_t N>
     using  static_ushort=boost::mpl::integral_c<ushort_t,N>;
 #else
-    typedef int                     int_t;
+    typedef long int                     int_t;
     typedef int                     short_t;
-    typedef int                     uint_t;
-    typedef int                     ushort_t;
+    typedef unsigned int                     uint_t;
+    typedef unsigned int                     ushort_t;
     template<int_t N>
     struct static_int : boost::mpl::integral_c<int_t,N>{
         typedef boost::mpl::integral_c<int_t,N> type;
